test(diet)：建立 CreateMeals 單元測試

+++ internal/dto/meal.go
@@ -12,10 +12,10 @@ type MealType struct {
 }
 
 type MealParamItem struct {
-	DietID int64 `json:"diet_id" gorm:"column:diet_id"` //飲食紀錄id
-	FoodID int64 `json:"food_id" gorm:"column:food_id"` //食物id
-	Type   int   `json:"type" gorm:"column:type"`       //類型(1:/早餐/2:午餐/3:晚餐/4:點心)
-	Amount int   `json:"amount" gorm:"column:amount"`   //數量
+	DietID int64   `json:"diet_id" gorm:"column:diet_id"` //飲食紀錄id
+	FoodID int64   `json:"food_id" gorm:"column:food_id"` //食物id
+	Type   int     `json:"type" gorm:"column:type"`       //類型(1:/早餐/2:午餐/3:晚餐/4:點心)
+	Amount float64 `json:"amount" gorm:"column:amount"`   //數量
 }
 
 type CreateMealsParam struct {
